build: add environment variables to test when api is off
