+++ .env
@@ -1,0 +1,2 @@
+REACT_APP_RANDOM_API=https://programming-quotesapi.vercel.app/api/random
+REACT_APP_ALTERNATIVE_API=https://southparkquotes.onrender.com/v1/quotes
