+++ .env.test
@@ -1,0 +1,2 @@
+REACT_APP_RANDOM_API=http://localhost:4000/random
+REACT_APP_ALTERNATIVE_API=http://localhost:4000/alternative
