+++ README.md
@@ -1,10 +1,22 @@
+<div align="center">
+
 # Genus Teach Post
+![GitHub Tag](https://img.shields.io/github/v/tag/matheusmfranco/genus-teach-post) ![](https://img.shields.io/badge/Framework-react-brightgreen) [![Basic validation](https://github.com/actions/labeler/actions/workflows/basic-validation.yml/badge.svg?branch=main)](https://github.com/actions/labeler/actions/workflows/basic-validation.yml)
+
+</div>
 
 A simple React application built with TypeScript that generates random programming quotes to inspire reflection and calmness.
 
+## Preview 
+### Mobile
+![Desktop](/prints/mobile.png)
+
+### Desktop
+![Desktop](/prints/desktop.png)
+
 ## Features
 
-- Displays random programming quotes;
+- Displays random programming quotes every 10 minutes;
 - Built with React and TypeScript for a type-safe, efficient, and reactive experience;
 - Responsive and user-friendly design for mobile and desktop.
 
@@ -14,28 +26,20 @@ A simple React application built with TypeScript that generates random programmi
 - **TypeScript**: Type-safe JavaScript with enhanced productivity and readability;
 - **LESS**: CSS pre-processor styling for an appealing design.
 
-## Installation
+## Local
 
 ```bash
 npm install
-```
-
-or
-
-```bash
-docker build -t genus-teach-post .
-```
-
-## Usage
-
-To start the application, run:
-
-```bash
 npm start
 ```
 
-or
+## Docker
 
 ```bash
-docker run -d -p 3000:80 genus-teach-post
+docker run -p 3000:80 matheusmagal/genus-teach-post
+```
+
+## Run Mock
+```bash
+npx json-server --watch db.json --port 4000
 ```
