+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "genus-teach-post",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
@@ -17,6 +17,7 @@
   },
   "scripts": {
     "start": "craco start",
+    "start:dev": "cross-env NODE_ENV=test npm start",
     "build": "craco build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -41,6 +42,7 @@
   },
   "devDependencies": {
     "@craco/craco": "^7.1.0",
+    "cross-env": "^7.0.3",
     "typescript": "^4.9.5"
   }
 }
